refactor: result 페이지 영수증 포맷 유틸을 _utils 로 추출
