+++ apps/web/src/app/tournament/[id]/result/_components/ReceiptPaper.tsx
@@ -8,6 +8,7 @@ import ReceiptZigzag from '@/assets/images/tournament/result/receipt-zigzag.svg'
 import { cn } from '@/utils/cn';
 
 import type { RankedProductT } from '../../_common/_types/tournament';
+import { formatDate, formatPrice, formatTime } from '../_utils/formatReceipt';
 
 const kodeMono = Kode_Mono({ subsets: ['latin'], weight: ['400', '500', '600', '700'] });
 
@@ -23,21 +24,6 @@ type ReceiptPaperProps = {
   canSharePlayLink: boolean;
   onSharePlayLink?: () => void;
 };
-
-const formatDate = (date: Date) => {
-  const dd = String(date.getDate()).padStart(2, '0');
-  const mm = String(date.getMonth() + 1).padStart(2, '0');
-  const yyyy = date.getFullYear();
-  return `${dd}/${mm}/${yyyy}`;
-};
-
-const formatTime = (date: Date) => {
-  const hh = String(date.getHours()).padStart(2, '0');
-  const mm = String(date.getMinutes()).padStart(2, '0');
-  return `${hh}:${mm}`;
-};
-
-const formatPrice = (price: number) => `${price.toLocaleString('ko-KR')}원`;
 
 const SectionDivider = () => <div className="h-px w-full border-t border-dashed border-gray-100" />;
