+++ apps/web/src/app/tournament/[id]/result/_utils/formatReceipt.ts
@@ -1,0 +1,14 @@
+export const formatDate = (date: Date) => {
+  const dd = String(date.getDate()).padStart(2, '0');
+  const mm = String(date.getMonth() + 1).padStart(2, '0');
+  const yyyy = date.getFullYear();
+  return `${dd}/${mm}/${yyyy}`;
+};
+
+export const formatTime = (date: Date) => {
+  const hh = String(date.getHours()).padStart(2, '0');
+  const mm = String(date.getMinutes()).padStart(2, '0');
+  return `${hh}:${mm}`;
+};
+
+export const formatPrice = (price: number) => `${price.toLocaleString('ko-KR')}원`;
